refactor: :truck: Rename components
