--- src/lib/components/atoms/PageTransition.svelte
@@ -1,15 +1,0 @@
-<script>
-  import { fly } from "svelte/transition";
-
-  export let url;
-</script>
-
-{#key url}
-  <div
-    in:fly={{ y: -50, duration: 250, delay: 300 }}
-    out:fly={{ y: -50, duration: 250 }} 
-  >
-    <slot/>
-  </div>
-{/key}
-
